% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRS_function.R
\name{BRS}
\alias{BRS}
\title{Run BRS}
\usage{
BRS(
  df,
  Y,
  maxLen,
  trainProp,
  numIter = 500L,
  numChain = 2L,
  numMine = 5000L,
  supp = 5L,
  alpha_1 = 50L,
  beta_1 = 1L,
  alpha_2 = 50L,
  beta_2 = 1L,
  prior_type = "beta",
  alpha_l = NULL,
  beta_l = NULL,
  lambda = NULL,
  nu = NULL,
  print = FALSE,
  bootstrap = FALSE,
  reps = 1L,
  sampleSize = NULL,
  seed = NULL
)
}
\arguments{
\item{df}{dataframe of binary features}

\item{Y}{vector of binary outcome}

\item{maxLen}{integer maximum length of rules}

\item{trainProp}{numeric for proportion of data to use as training data}

\item{numIter}{integer for number of iterations in simulated annealing}

\item{numChain}{integer for number of chains of simulated annealing}

\item{numMine}{integer for number of rules to mine}

\item{supp}{integer for percent minimum support}

\item{alpha_1}{numeric for alpha_+ from the paper}

\item{beta_1}{beta_+}

\item{alpha_2}{alpha_-}

\item{beta_2}{beta_-}

\item{prior_type}{string for the prior type. Either "beta" for BRS-BetaBinomial or "poisson" for BRS-Poisson}

\item{alpha_l}{vector of alpha_l for l=1...maxLen. If set to NULL and prior_type="beta", values will be automatically generated. Ignored if prior_type="poisson"}

\item{beta_l}{vector of beta_l for l=1...maxLen. If set to NULL and prior_type="beta", values will be automatically generated. Ignored if prior_type="poisson"}

\item{lambda}{numeric rate parameter for the prior on the number of rules}

\item{nu}{numeric rate parameter for the prior on the length of rules}

\item{print}{logical whether to print progress of algorithm}

\item{bootstrap}{logical for whether to bootstrap}

\item{reps}{integer for number of bootstrap reps}

\item{sampleSize}{integer for bootstrap sample size. If set to NULL, default is the number of observations in data}
}
\value{
list of rule sets from all bootstrap replications

indices of bootstrap samples

accuracy, true positive rate, and false positive rate on test data for each rule set
}
\description{
This function runs BRS. It allows the user either to run BRS only once on the
original data or to run BRS on bootstrapped samples
}
