% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chord.R
\name{.get_df_chord}
\alias{.get_df_chord}
\title{Make data frame for circilize}
\usage{
.get_df_chord(allRuleSets, featureGroups, maxLen, minProp)
}
\value{
data frame
}
\description{
Make data frame for circilize
}
