% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{getLabel}
\alias{getLabel}
\title{Get labels for features}
\usage{
getLabel(feat, labels_df, neg_label)
}
\arguments{
\item{feat}{feature or vector of features to get the label of; should be of the form "feature_value"}

\item{labels_df}{dataframe of unique feature and corresponding labels; first column is the feature, second column is corresponding label}

\item{neg_label}{prefix to use for negative features}
}
\value{
label to use for feat
}
\description{
Get the labels of a features to use in graphs
}
