% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{numCases}
\alias{numCases}
\title{Get the number of cases that satisfy a rule}
\usage{
numCases(rule, df, Y = NULL, y_val = 1)
}
\arguments{
\item{rule}{rule to check: must be in the format feature_value or feature_value_neg}

\item{df}{data to check}

\item{Y}{if included, checks that case's Y=y_val}

\item{y_val}{value Y should take}
}
\value{
number of cases in data that satisfies rule
}
\description{
Get the number of cases that satisfy a rule
}
