% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar.R
\name{plot_bar}
\alias{plot_bar}
\title{Make bar plot}
\usage{
plot_bar(
  df,
  Y,
  fit,
  featureLabels,
  maxLen,
  topRules = 10,
  and = " AND ",
  neg = "NOT ",
  minProp = 0,
  simplify = T,
  oppmat = NULL,
  oppind = NULL,
  heightBuffer = 1,
  plotBuffer = 0,
  textSize = 16,
  boot_rep = 100
)
}
\arguments{
\item{df}{dataframe of binary features}

\item{Y}{vector of binary outcome}

\item{fit}{output from function \code{BRS}}

\item{featureLabels}{data frame whose first column is feature names (as they appear in df) and whose second column is the corresponding labels to be displayed on the graph}

\item{maxLen}{integer maximum length of rules}

\item{topRules}{the max number of rules to plot for each length}

\item{and}{string for how an "and" statement will be concatenated}

\item{neg}{string for how a "not" statement will be prefaced}

\item{minProp}{minimum proportion of rule sets in which a rule must appear before being plotted}

\item{simplify}{whether to simplify negative rules when possible. If true, then a negative rule is changed to it's positive counterpart if and only if there are only two possible values for the corresponding variable (e.g. "A_1_neg" is changed to "A_0" if and only if the only two possible values for "A" are 0 and 1)}

\item{heightBuffer}{amount by which to shift bottom of graph up to make room for label}

\item{textSize}{\code{size} parameter passed to \code{ggplot2::element_text}}
}
\value{
bar plot with frequency and coverage of rules
}
\description{
This function makes a bar plot of rule frequency and coverage
}
