% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsne.R
\name{plot_tsne}
\alias{plot_tsne}
\title{Make a t-SNE plot}
\usage{
plot_tsne(
  df,
  Y,
  A,
  caseColors,
  boxColor,
  pointSize = 1,
  textSize = 1,
  symb = 19,
  bottom_buffer = 1.25,
  all_buffer = 1,
  legend_under_plot = T,
  legend_bg_col = "transparent",
  legend_offset = c(0, 0),
  legend_position = "bottomright",
  shade_name = "Positive classification",
  jitter_factor = 1,
  jitter_amount = NULL,
  max_iter = 1000
)
}
\arguments{
\item{df}{dataframe of binary features}

\item{Y}{vector of binary outcome}

\item{A}{rule set to use for determining which variables will be used for training t-SNE (will use outcome plus all variables that appear in \code{A})}

\item{caseColors}{vector colors to use for classification correctness. The first element is for correct, second is for incorrect}

\item{boxColor}{color for shading}

\item{pointSize}{\code{cex} paramater for \code{plot} for size of points}

\item{textSize}{\code{cex} paramater for \code{legend} for size of text}

\item{symb}{\code{pch} paramater for \code{plot} for symbol of points}

\item{bottom_buffer}{amount by which to shift bottom of graph up for legend}

\item{all_buffer}{buffer on all sides of plot}

\item{legend_under_plot}{whether legend is under plot or inside the plot. Overrides \code{legend_coordinates}}

\item{legend_bg_col}{background color of legend}

\item{legend_offset}{distance to move legend (as a percentage). Negative values move legend to the left and down, positive values move legend to the right and up. If \code{legend_under_plot=F}, then this argument is analagous to the inset arguement for \code{\link[graphics]{plot}}}

\item{legend_position}{position of legend}

\item{shade_name}{label in legend for shaded points}

\item{jitter_factor}{factor by which to jitter points}

\item{jitter_amount}{amount by which to jitter points}

\item{max_iter}{maximum number of iterations to run t-SNE}
}
\value{
t-SNE plot
}
\description{
Makes a t-SNE plot with color coding for classification correctness and shading for predicted outcome
}
