% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{simplifyFeature}
\alias{simplifyFeature}
\title{Function that simplifies feature if possible}
\usage{
simplifyFeature(originalFeature, featureNames, featureValues)
}
\arguments{
\item{originalFeature}{feature to simplify}

\item{featureNames}{vector of all feature names, repeated according to multiplicity}

\item{featureValues}{vector of all values corresponding to features (same order as featureNames)}
}
\value{
name of feature to use
Original feature if no _neg suffix
if _neg suffix: original feature and value if none or multiple alternative values, other feature value if exactly one alternative
}
\description{
Function that simplifies feature if possible
}
